requirements-dev.txt
-r requirements-test.txt
black
isort
toml
flake8
nbstripout
jupytext
radvel
rebound
rebound_pymc3>=0.0.3
jupyterlab
lightkurve
